refactor to support auto ollama serve
