+++ utils/model.py
@@ -77,41 +77,7 @@ def load_model(args: SakuraModelConfig):
         from llama_cpp import Llama
 
     if args.ollama:
-        import ollama
-        from tqdm import tqdm
-
-        class Ollama:
-            '''Llama-style wrapper for ollama'''
-            def __init__(self, model):
-                self.model = model
-                self.pull()
-
-            def __call__(self, prompt, stream=False, **kwargs):
-                return ollama.generate(self.model, prompt, stream=stream, options=kwargs)
-
-            def pull(self):
-                current_digest, bars = "", {}
-                for progress in ollama.pull(self.model, stream=True):
-                    digest = progress.get("digest", "")
-                    if digest != current_digest and current_digest in bars:
-                        bars[current_digest].close()
-
-                    if not digest:
-                        print(progress.get("status"))
-                        continue
-
-                    if digest not in bars and (total := progress.get("total")):
-                        bars[digest] = tqdm(
-                            total=total,
-                            desc=f"pulling {digest[7:19]}",
-                            unit="B",
-                            unit_scale=True,
-                        )
-
-                    if completed := progress.get("completed"):
-                        bars[digest].update(completed - bars[digest].n)
-
-                    current_digest = digest
+        from .ollama import Ollama
 
     if args.vllm:
         from vllm import AsyncEngineArgs, AsyncLLMEngine, LLM
@@ -425,7 +391,7 @@ class SakuraModel:
     ):
         output = model(
             prompt,
-            num_ctx=generation_config.__dict__["max_new_tokens"],
+            num_predict=generation_config.__dict__["max_new_tokens"],
             temperature=generation_config.__dict__["temperature"],
             top_p=generation_config.__dict__["top_p"],
             repeat_penalty=generation_config.__dict__["repetition_penalty"],
@@ -447,7 +413,7 @@ class SakuraModel:
         for output in model(
             prompt,
             stream=True,
-            num_ctx=generation_config.__dict__["max_new_tokens"],
+            num_predict=generation_config.__dict__["max_new_tokens"],
             temperature=generation_config.__dict__["temperature"],
             top_p=generation_config.__dict__["top_p"],
             repeat_penalty=generation_config.__dict__["repetition_penalty"],
